% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{as_probability}
\alias{as_probability}
\title{Function for forcing probabilities between zero and one}
\usage{
as_probability(p)
}
\arguments{
\item{p}{A real number}
}
\value{
A real number between 0 and 1
}
\description{
Function for forcing probabilities between zero and one
}
