% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEV_functions.R
\name{buildformula_factors}
\alias{buildformula_factors}
\title{list of non-null input parameters}
\usage{
buildformula_factors(
  binary_code_dataframe,
  FACTOR_NAMES_DATAFRAME,
  list_of_factors = NULL
)
}
\description{
list of non-null input parameters
}
