% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{check_arguments}
\alias{check_arguments}
\title{Check that input arguments are correct}
\usage{
check_arguments(resp_var, predictors)
}
\arguments{
\item{resp_var}{(str) one of 'fev1' or 'fev1_fvc'}
}
\description{
Check that input arguments are correct
}
