% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_model.R
\name{create_lmer_model}
\alias{create_lmer_model}
\title{Create LMER Model}
\usage{
create_lmer_model(resp_var, lmer_model_file, lmer_model_file_binary, data_dir)
}
\arguments{
\item{resp_var}{(str) one of 'fev1' or 'fev1_fvc'}

\item{lmer_model_file}{(str) full file name containing the Linear Mixed Effects model}

\item{lmer_model_file_binary}{(str) shortened file name, containing only 1's and 0's,
where each 1 or 0 corresponds to the presence/absence of one of the predictor variables.}

\item{data_dir}{(str) full directory name where the data for the package is located}
}
\description{
Create LMER Model
}
