% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_model.R
\name{lmer_model_exists}
\alias{lmer_model_exists}
\title{Check if the Linear Mixed Model file has been created already}
\usage{
lmer_model_exists(lmer_model_file)
}
\arguments{
\item{lmer_model_file}{(str) full name of lmer model file}
}
\description{
Check if the Linear Mixed Model file has been created already
}
