% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{make_predictions}
\alias{make_predictions}
\title{Generate lung function predictions}
\usage{
make_predictions(resp_var, predictors, lmer_model_file = NA)
}
\arguments{
\item{resp_var}{(str) one of 'fev1' or 'fev1_fvc'}

\item{predictors}{(data.frame) a data frame containing the following columns:

\itemize{
\item fev1_0. (float) (REQUIRED) forced expiratory volume in 1 second in L (0--5)
\item fvc_0. (float) (REQUIRED) force vital capacity, the amount of air that can be forcibly
blown out in one exhale (L)
\item age. (float) (REQUIRED) the age of the patient (20--100)
\item triglycerides. (float or NA) blood triglyceride levels (mg/dL)
\item hematocrit. (float or NA) blood hematocrit percentage
\item albumin. (float or NA) blood albumin (g/L)
\item globulin. (float or NA) blood globulin (g/L)
\item ALP. (float or NA) blood alkaline phosphotase (IU/L)
\item WBC. (float or NA) white blood cells (10^9/L)
\item QRS_intv. (float or NA) TODO
\item wine. (float or NA) number of glasses of wine consumed per week
\item cocktail. (float or NA) TODO
\item beer. (float or NA) number of beers (cans/bottles) consumed per week
\item height. (float or NA) (REQUIRED) patient height in centimetres (100--250)
\item smoke_year. (float or NA) number of years the patient has been smoking for
\item daily_cigs. (float or NA) number of cigarettes the patient smokes per day
\item sex. (str) (REQUIRED) sex of patient, one of 'male', 'female'
\item broncho. (str) TODO
\item dyspnea_exc. (str) TODO
\item night_symp. (str) TODO
}}

\item{lmer_model_file}{(str) name of file with linear mixed model fit by maximum likelihood, from
lmer package \url{https://www.rdocumentation.org/packages/lme4/versions/1.1-23/topics/lmer}}
}
\value{
data_pred_fin (data.frame) a data frame containing predictions:

\itemize{
\item year (int)
\item predicted_FEV1
\item predicted_SE_PI
\item lowerbound_PI
\item upperbound_PI
\item predicted_SE_CI
\item percentpred
\item percentpred_upperbound_PI
\item percentpred_lowerbound_PI
\item predicted_FEV1_if_smoke
\item FEV1_lowerbound_CI_if_smoke
\item upperbound_CI_if_smoke
\item predicted_FEV1_if_quit
\item FEV1_lowerbound_CI_if_quit
\item upperbound_CI_if_quit
\item cpackyr_if_smoke
\item cpackyr_if_quit
\item percentpred_if_smoke
\item percentpred_FEV1_lowerbound_CI_if_smoke
\item percentpred_upperbound_CI_if_smoke
\item percentpred_if_quit
\item percentpred_FEV1_lowerbound_CI_if_quit
\item percentpred_upperbound_CI_if_quit
}
}
\description{
Generate lung function predictions
}
