% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-init.R
\docType{data}
\name{sample_predictors}
\alias{sample_predictors}
\title{Sample Patient Characteristics Inputs}
\format{
A data frame with 1 row and 20 variables
}
\description{
A dataset containing sample patient characteristics to run the prediction model
variables are as follows:
}
\details{
\itemize{
\item fev1_0. (float) (REQUIRED) forced expiratory volume in 1 second in L (0--5)
\item fvc_0. (float) (REQUIRED) TODO
\item age. (float) (REQUIRED) the age of the patient (20--100)
\item triglycerides. (float or NA) blood triglyceride levels (mg/dL)
\item hematocrit. (float or NA) blood hematocrit percentage
\item albumin. (float or NA) blood albumin (g/L)
\item globulin. (float or NA) blood globulin (g/L)
\item ALP. (float or NA) blood alkaline phosphotase (IU/L)
\item WBC. (float or NA) white blood cells (10^9/L)
\item QRS_intv. (float or NA) TODO
\item wine. (float or NA) number of glasses of wine consumed per week
\item cocktail. (float or NA) TODO
\item beer. (float or NA) number of beers (cans/bottles) consumed per week
\item height. (float or NA) (REQUIRED) patient height in centimetres (100--250)
\item smoke_year. (float or NA) number of years the patient has been smoking for
\item daily_cigs. (float or NA) number of cigarettes the patient smokes per day
\item sex. (str) (REQUIRED) sex of patient, one of 'male', 'female'
\item broncho. (str) TODO
\item dyspnea_exc. (str) TODO
\item night_symp. (str) TODO
}
}
\keyword{datasets}
